JS API: Refactor the code and split into multiple modules.
